Added summernote-ext-emoji-ajax
All the other emoji extensions I tried were either old/broken or weren't doing it for me. Mostly I had problems with the browser stalling as it loaded the images, and the styling didn't match the rest of summernote.

This extension uses the github emoji source, similar to the Summernote Emoji Hint example, except it uses AJAX asynch.